Moved Ceres math utilities to utils directory
