+++ ical_core/include/ical_core/cost_functions/camera_intrinsic.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <ical_core/types.h>
-#include <ical_core/ceres_math_utilities.h>
+#include <ical_core/optimizations/utils/ceres_math_utilities.h>
 
 namespace industrial_calibration
 {
