+++ ical_core/include/ical_core/optimizations/dh_chain_kinematic_calibration.h
@@ -2,7 +2,7 @@
 
 #include <ical_core/types.h>
 #include <ical_core/dh_chain.h>
-#include <ical_core/ceres_math_utilities.h>
+#include <ical_core/optimizations/utils/ceres_math_utilities.h>
 #include <ical_core/optimizations/utils/covariance_analysis.h>
 
 namespace industrial_calibration
