+++ ical_core/include/ical_core/optimizations/utils/ceres_math_utilities.h
