+++ ical_core/src/optimizations/utils/ceres_math_utilities.cpp
@@ -1,4 +1,4 @@
-#include <ical_core/ceres_math_utilities.h>
+#include <ical_core/optimizations/utils/ceres_math_utilities.h>
 
 namespace industrial_calibration
 {
